added the features for ban functionality, removes JS for old version
