+++ lms/urls.py
@@ -199,6 +199,10 @@ if settings.WIKI_ENABLED:
 
         # Include edly panel app URL's
         url(r'^api/edly_panel/', include('edly_panel_app.api.urls', namespace='edly_panel_api')),
+
+        # Inlcude lumsx feature app urls
+        url(r'^lumsx_features/', include('openedx.features.lumsx_features.urls', namespace='lumsx_features')),
+
     ]
 
 COURSE_URLS = [
